% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEBpath-package.R
\docType{package}
\name{DEEBpath-package}
\alias{DEEBpath}
\alias{DEEBpath-package}
\title{DEEBpath: Handle Paths, Folders, Filenames for the DEEB Project}
\description{
Handle paths, folders, filenames for the DEEB project.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
